Backend and REFR UI fixes; Sustain Phase 1 plans
Qt chokes if you use Q_OBJECT on a nested class definition. We don't actually need it here, so we'll ditch it.

Constexpr and template-inclusion fixes within the backend.

Fixed handling of ExtraAlphaCutoff, including making it possible for the user to explicitly force it to its default via a checkbox.

I misunderstood what std::mem_fn does; replaced it with a custom template that has the desired effect, and a macro that allows for better ergonomics than C++'s syntax rules re: pointer-to-member-functions do.

With this commit, the program now builds again. The REFR UI still needs testing.