Main window MDI fixes + DKVulkanView improvements
Implemented some optimizations for DKVulkanView. However, I can't repro the lag I saw on the Render Window back when I decided to have it closed by default, so out of an abundance of caution I am leaving it closed by default. Also improved the widget to use QBasicTimer instead of handling timer IDs by hand.

Fixed a bug where you have to select "Render Window" in the "Windows" menu twice in order to display it.

Fixed a bug where closing and reopening the Log Window caused it to reopen with its contents not visible, due to a mistake in QMdiSubWindow's implementation.

Resizing the Render Window now takes advantage of QMdiSubWindow::SubWindowOption::RubberBandResize. This means that instead of resizing the window in real-time (which causes "hall of mirrors" ghosting), Qt instead shows a resize rectangle without repainting the contents of the window... or so the documentation claims. In practice, Qt is displaying a blue-tinted rectangle, which ends up having a shimmer effect because the position is randomly inaccurate by one pixel on any given frame (and the visual design doesn't even match what's native for Windows anyway, so it's double cool -_-). I'm keeping it anyway. It's late; I'm tired; it's stayin' in.